// templates/graphql-ts-app/src/types/user.ts
export interface User {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface CreateUserInput {
  name: string;
  email: string;
}

export interface UpdateUserInput {
  id: string;
  name?: string;
  email?: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}